# entertainmentEmporiumBackend/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Entertainment Emporium Backend API 
  description: A movies and tv shows rating API 
  contact:
    name: Francesca 
    email: passmoref@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://hexagonprofile-maximumalex-3000.codio-box.uk/api/v1'
    description: Development server
...
paths:
  /users/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the user
    get:
      tags:
        - Users
      description:
        Get user by id.
      responses:
        '200':
          description: View of a user’s record, excluding password.
        '404':
          description: User was not found in the DB.
        '403':
          description: Requester forbidden to view this resource
    put:
      tags:
        - Users
      description:
        Update user by ID.
      responses:
        '200':
          description: Update confirmation and link to record.
        '404':
          description: User was not found in the DB.
        '401':
          description: Users can only update their own accounts
        '403':
          description: Requester forbidden to update this resource

    delete:
      tags:
        - Users
      description:
        Delete user by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only admin users can delete user accounts.
        '404':
          description: User was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /users:
    summary: Represents an individual user
    description: >
      This resource represents an individual user in the system.
      Each user is identified by a numeric `id`.
    post:
      tags:
        - Users
      description:
        Create a new user account.
      requestBody:
        description: All new user data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./users.json#/definitions/user
      responses:
        '200':
          description: User created
          content:
            application/json:
              schema:
                $ref: ./users.json#/definitions/userCreated
  /users/{username}:
    parameters:
      - in: path
        name: username
        schema:
          type: string
        required: true
        description: The unique username of the user
    get:
      tags:
        - Users
      description:
        Get user by username.
      responses:
        '200':
          description: View of a user’s record, excluding password.
        '404':
          description: User was not found in the DB.
        '403':
          description: Requester forbidden to view this resource
...